Refactored to pods and core structure
