+++ back/.babelrc
@@ -22,8 +22,6 @@
                       "core": "./src/core",
                       "dals": "./src/dals",
                       "pods": "./src/pods",
-                      "messages": "./src/messages",
-                      "business": "./src/business"
                     }
                   }
               ]
