+++ back/.env
@@ -1,7 +1,7 @@
 NODE_ENV=development
 PORT=8081
-SOCKET_PORT=3000
 MOCK_REPOSITORY=false
 API_URL=/api
 CORS_ORIGIN=http://localhost:8080
 MONGODB_URI=mongodb://localhost:27017/code-paster
+STATIC_FILES_PATH="../../front/dist"
