--- back/src/business/index.ts
@@ -1,1 +1,0 @@
-export * from './room-name-generator.business';
