+++ back/src/core/constants/header.constants.ts
@@ -1,0 +1,5 @@
+export const headerConstants = {
+    protocol: 'x-forwarded-proto',
+    httpsProtocol: 'https',
+    host: 'host',
+  };
