+++ back/src/core/constants/index.ts
@@ -1,1 +1,2 @@
 export * from './env.constants';
+export * from './header.constants';
