+++ back/src/core/servers/express.server.ts
@@ -1,6 +1,8 @@
 import express from 'express';
 import bodyParser from 'body-parser';
 import { corsOptions } from './cors'
+import { envConstants } from 'core/constants';
+import { redirectHttpsMiddleware } from './middlewares';
 
 export const createApp = () => {
   const app = express();
@@ -8,5 +10,10 @@ export const createApp = () => {
   app.use(bodyParser.urlencoded({ extended: false }));
   app.use(bodyParser.json());
   app.use(corsOptions);
+
+  if (envConstants.isProduction) {
+    app.use(redirectHttpsMiddleware);
+  }
+
   return app;
 };
