+++ back/src/core/servers/index.ts
@@ -1,4 +1,2 @@
 export * from './express.server';
-
-//TODO: refactoring
-//export * from './sockets.server';
+export * from './sockets.server'
