+++ back/src/core/servers/middlewares/index.ts
@@ -1,0 +1,1 @@
+export * from './redirect-https.middleware';
