+++ front/dev.env
@@ -1,3 +1,3 @@
 NODE_ENV=development
 BASE_API_URL=http://localhost:8081
-BASE_SOCKET_URL=http://localhost:3001
+BASE_SOCKET_URL=http://localhost:8081
